lydian-api/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
MarkupSafe==3.0.2
openai==1.54.4
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.11.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
